Wine glasses detection
Put wine glasses in a nine-box grid and output a 3x3 matrix. The case false, null means that there is no wine glass in the square; the case true, transparent means that there is an empty wine glass in the square; the case true, red means that there is a glass with red wine in the square.